flatting file structure
